pyproject.toml
[project]
name = "fastapi-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[standard]>=0.115.12",
    "langchain-core>=0.3.67",
    "langchain[openai]>=0.3.26",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langgraph[postgres]>=0.5.1",
    "openai>=1.93.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]
